removed pypesvds
The pylons web framework is no longer developed in python 3. The whole thing
would have to be ported to py3k and pyramid. This is a substantial amount of
work, just to have an admittedly nice GUI.

It would also be useful for a cluster deployment of a pypes network, but
I don't have the time to do this conversion now, as most of it would have to
be rewritten.